@@ -18,12 +18,10 @@
     "query-string": "5",
     "rc-queue-anim": "^1.6.12",
     "rc-tween-one": "^2.4.1",
-    "react": "^16.3.2",
+    "react": "16.8.6",
     "react-burger-menu": "^2.6.10",
     "react-clipboard.js": "^2.0.7",
-    "react-device-detect": "^1.7.5",
-    "react-dom": "^16.3.2",
-    "react-draggable": "^3.3.0",
+    "react-dom": "16.8.6",
     "react-dropzone": "4.2.10",
     "react-identicons": "^1.1.7",
     "uuid": "^3.3.2"
@@ -54,7 +52,7 @@
     "react-app-rewire-less": "^2.1.1",
     "react-app-rewire-yaml": "^1.1.0",
     "react-app-rewired": "^1.5.2",
-    "react-scripts-ts": "^2.16.0",
+    "react-scripts-ts": "^3.1.0",
     "rimraf": "^2.6.3",
     "ts-import-plugin": "^1.5.0",
     "tslint": "^5.9.1",
