@@ -3,7 +3,7 @@ import * as React from "react";
 import RealtimeRoom from "./RealtimeRoom";
 import message from "antd/lib/message";
 
-import {JoinRoomParams, PptConverter, Room, RoomPhase, RoomState, WhiteWebSdk, AnimationMode} from "white-web-sdk";
+import {JoinRoomParams, PptConverter, Room, RoomPhase, RoomState, WhiteWebSdk, AnimationMode, ViewMode} from "white-web-sdk";
 import {UserPayload} from "../common";
 import {UserCursor, LoadingPage} from "../components";
 import {OSSOptions} from "../tools";
@@ -14,6 +14,7 @@ export type RealtimeRoomPageProps = {
     readonly userPayload: UserPayload;
     readonly sdk: WhiteWebSdk;
     readonly ossOptions: OSSOptions;
+    readonly isWritable?: boolean;
     readonly disableAppFeatures?: boolean;
     readonly callbacks?: RealtimeRoomPageCallbacks;
 };
@@ -65,6 +66,7 @@ export default class RealtimeRoomPage extends React.Component<RealtimeRoomPagePr
             roomToken: this.roomToken,
             cursorAdapter: userCursor,
             userPayload: {...this.props.userPayload},
+            isWritable: this.props.isWritable,
         };
         const room = await whiteWebSdk.joinRoom(roomParams, {
             onPhaseChanged: phase => {
@@ -86,15 +88,16 @@ export default class RealtimeRoomPage extends React.Component<RealtimeRoomPagePr
         });
         userCursor.refreshRoomMembers(room.state.roomMembers);
 
-        // 对准 ppt
-        room.moveCameraToContain({
-            originX: -640,
-            originY: -360,
-            width: 1280,
-            height: 720,
-            animationMode: AnimationMode.Immediately,
-        });
-
+        if (room.state.broadcastState.mode !== ViewMode.Follower) {
+            // 对准 ppt
+            room.moveCameraToContain({
+                originX: -640,
+                originY: -360,
+                width: 1280,
+                height: 720,
+                animationMode: AnimationMode.Immediately,
+            });
+        }
         (window as any).room = room;
         this.setState({
             room: room,
